pyproject.toml
[project]
name = "promise-service-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[all]>=0.115.6,<0.116.0",
    "sqlmodel>=0.0.22,<0.1.0",
    "sqlalchemy-utc>=0.14.0,<0.15.0",
    "aiosqlite>=0.20.0,<0.21.0",
    "alembic>=1.14.0,<1.15.0",
    "greenlet>=3.1.1,<3.2.0",
]

[dependency-groups]
dev = [
    "black>=26.3.1",
    "pytest>=8.3.4,<9.0.0",
    "pytest-asyncio>=0.24.0,<1.0.0",
]

[tool.pytest.ini_options]
minversion = "8.3.4"
env = []
addopts = " --strict-markers --tb=short --asyncio-mode=auto -p no:warnings --doctest-modules"
python_files = ["tests.py", "test_*.py"]
rootdir = "./"
testpaths = ["./tests", "./appserver"]
pythonpath = ["./appserver"]
filterwarnings = ["error", "ignore::DeprecationWarning:etcd3.*:"]
log_cli = true
log_cli_level = "WARNING"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
